src/requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.6
django-dotenv==1.4.2
django-extensions==2.2.9
django-phonenumber-field==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
phonenumberslite==8.12.3
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
